// prisma/graphql/generated/graphql/resolvers/inputs/DoctorsWhereInput.ts
export class DoctorsWhereInput {
  @TypeGraphQL.Field(_type => [DoctorsWhereInput], {
    nullable: true
  })
  AND?: DoctorsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [DoctorsWhereInput], {
    nullable: true
  })
  OR?: DoctorsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [DoctorsWhereInput], {
    nullable: true
  })
  NOT?: DoctorsWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => UuidFilter, {
    nullable: true
  })
  id?: UuidFilter | undefined;

  @TypeGraphQL.Field(_type => Enumdoctors_type_enumFilter, {
    nullable: true
  })
  type?: Enumdoctors_type_enumFilter | undefined;

  @TypeGraphQL.Field(_type => BoolFilter, {
    nullable: true
  })
  is_license_verified?: BoolFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  bio?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableListFilter, {
    nullable: true
  })
  education?: StringNullableListFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableListFilter, {
    nullable: true
  })
  experience?: StringNullableListFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  first_name?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableListFilter, {
    nullable: true
  })
  languages?: StringNullableListFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  last_name?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  profile_image?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => Enumdoctors_specialty_enumFilter, {
    nullable: true
  })
  specialty?: Enumdoctors_specialty_enumFilter | undefined;

  @TypeGraphQL.Field(_type => UuidFilter, {
    nullable: true
  })
  user_id?: UuidFilter | undefined;

  @TypeGraphQL.Field(_type => CertificatesListRelationFilter, {
    nullable: true
  })
  certificates?: CertificatesListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => ConsultationsListRelationFilter, {
    nullable: true
  })
  consultations?: ConsultationsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Doctor_institutsListRelationFilter, {
    nullable: true
  })
  doctor_instituts?: Doctor_institutsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => UsersRelationFilter, {
    nullable: true
  })
  users?: UsersRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Lab_requestsListRelationFilter, {
    nullable: true
  })
  lab_requests?: Lab_requestsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PrescriptionsListRelationFilter, {
    nullable: true
  })
  prescriptions?: PrescriptionsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Rdv_requestsListRelationFilter, {
    nullable: true
  })
  rdv_requests?: Rdv_requestsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => RdvsListRelationFilter, {
    nullable: true
  })
  rdvs?: RdvsListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => MessageListRelationFilter, {
    nullable: true
  })
  sentMessages?: MessageListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => MessageListRelationFilter, {
    nullable: true
  })
  receivedMessages?: MessageListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Instituts_medicauxListRelationFilter, {
    nullable: true
  })
  instituts?: Instituts_medicauxListRelationFilter | undefined;
}